backend/requirements.txt
fastapi==0.115.2
pydantic==2.9.2
python-dotenv==1.0.1
SQLAlchemy==2.0.36
uvicorn==0.32.0
